nlp-service/nlp/requirements.txt
spacy
rapidfuzz
requests>=2.25.0
urllib3>=1.26.0
python-dotenv>=0.19.0
certifi>=2021.10.8
fastapi
uvicorn[standard]
pymongo
typing-extensions
pydantic